refactor: app 폴더 layout 구조 변경 (#ATR-519)
