+++ apps/service/.eslintrc.cjs
@@ -6,6 +6,7 @@ module.exports = {
     tsconfigRootDir: __dirname,
   },
   rules: {
+    '@next/next/no-sync-scripts': 'off',
     'react/require-default-props': 'off',
     'jsx-a11y/label-has-associated-control': [
       2,
