+++ apps/service/app/(service)/(static)/discover/page.tsx
