+++ apps/service/app/(service)/(static)/inbox-bookmark/article/[articleId]/error.tsx
