+++ apps/service/app/(service)/(static)/inbox-bookmark/page.tsx
