+++ apps/service/app/(service)/(static)/inbox/article/[articleId]/page.tsx
