+++ apps/service/app/(service)/(static)/inbox/layout.tsx
