+++ apps/service/app/(service)/(static)/inbox/page.tsx
