+++ apps/service/app/(service)/(static)/mypage/page.tsx
