+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/article/[articleId]/loading.tsx
