+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/article/[articleId]/page.tsx
