+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/error.tsx
