+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/layout.tsx
