+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/loading.tsx
