+++ apps/service/app/(service)/(static)/newsletter/[newsletterId]/page.tsx
