+++ apps/service/app/(service)/(static)/offline/page.tsx
