+++ apps/service/app/(service)/(static)/page.tsx
@@ -1,7 +1,7 @@
 import { MainLayout } from '@/widgets/main'
 import { Header } from '@/widgets/menu'
 
-export default function Home() {
+export default function MainPage() {
   return (
     <>
       <Header title="홈" />
