+++ apps/service/app/(service)/(static)/policy/privacy/page.tsx
