+++ apps/service/app/(service)/(static)/setting/page.tsx
