+++ apps/service/app/(service)/global-error.tsx
@@ -16,9 +16,8 @@ export default function GlobalError({
   return (
     <html lang="ko">
       <body
-        className="relative min-h-dvh min-w-[280px] break-keep bg-gray-50 px-5 pb-40 text-gray-700 md:px-10 md:pt-10 dark:bg-gray-900 dark:text-gray-100"
+        className="service px-5 pb-40 md:px-10 md:pt-10"
         suppressHydrationWarning>
-        {/* eslint-disable-next-line @next/next/no-sync-scripts */}
         <script src="/script/theme.js" />
         {/* `NextError` is the default Next.js error page component. Its type
         definition requires a `statusCode` prop. However, since the App Router
